+++ src/arch/windows/process_handles.h
@@ -101,5 +101,6 @@ VOID ps__get_open_files_init();
 SEXP ps__get_open_files(long pid, HANDLE processHandle);
 DWORD ps__NtQueryObject(void);
 DWORD WINAPI ps__NtQueryObjectThread(LPVOID lpvParam);
+SEXP ps__get_modules(HANDLE hProcess);
 
 #endif // __PROCESS_HANDLES_H__